Various QoL changes
* Buffed the clairvoyance effect of the mask of insight from the shitty
3-tile-range to the 40-tile super-clairvoyance. Also made it drain
essense twice as fast. I really wish there was a middle ground between
"pretty much worthless" and "I've seen some shit"
* Reduced energy drain rate of the mantle of shadows by half.
* Changed weather effect of symbol of judgement to be on-wield, gave it
an effect on machines.
* Changed the hellfire stave's forced burst to instead be optional.
* Changed shot effect on wraithslayer to use "wide" instead.
* Gave demon claw a custom explosion instead of stacking an HE explosion
on a napalm one.
* Quality of life change: most professions (except apprentice) get at
least some access to a knife, firestarter, and container. Trivial to
improvise most of those innawoods, but it's largly just a brief bit of
tedium.
* Changed thunder gauntlet action from stungun effect to the electrical
artifact action.
* Small description change for symbol of judgement, sooomeone thought
from the description that the 5-volume lightning weapon was wearable.