The "Pierce the Veil" Update
* Adds a restored ritual blade that can be made from the cursed blade.
General ideas for the recipe thanks to
* Using the ritual blade can grant the new spell-like mutations with
enough essence.
* Activating a spell-like mutation consumes fatigue and grants an
inteligence debuff until the magic sign is activated, or until disposed
of in the case of most buff spells that don't rely on effects fuckery.
* Mutation category finally working complete with dreams and a
threshhold. You don't want to know how much experimenting it took me
before I figured out that it was as simple as using the "MUT_IV"
use_action instead.