Various requested updates
* Fixed potions freezing so readily, due using default freezing point.
* Added basic chemistry seat and chemistry set as options for rituals
using a bowl.
* Hopefully ensured summoned monsters that normally evolve can't.
* A few tweaks to a couple of the more underwhelming monsters (a status
effect for skeletons, ranged grab and lunge for the thing), and removed
tentacle attack as friendly monsters don't use it.
* Gave hit-and-run monsters improved vision range.
* Ensured blinds monsters can see, to aid targeting.
* Added to wraithslayer's effects, gaining minor stun and chance of
blindness.
* Implemented the feature where a magic seal can be crafted using a
magic sign instead of the monsterpart, for hoarding and/or better abuse
on NPCs.
* Added the first, easier half of the mutual exclusion: restored ritual
blade can no longer be used to craft the sacramental heart.
* Altered the intensify of scroll of darkness, being the only Six
Pillars scroll that was trivially rendered useless.