Various tweaks and fixes
* Updated dispersion of all magic ranged weapons to match how it works
in vanilla now.
* Tested and implemented armor_multilpier for magic ranged weapons, it
still won't show anything for armor-pierce but now wraithslayer goes
back to yeeting through armor as intended, and all others have their own
minor armor-penetration. Description also hints at this.
* Since arcae-skill ranged magic weapons can't accept rifle or crossbow
gunmods, it was suggested I simply make the sling part of the recipe.
Did so. The alternative would've been to make it use rifles skill
(fitting as it's the only one to handle like a mundane ranged weapon),
but this was a good alternative too (as grips aren't of much use to a
single-shot weapon with low recoil, variant sights are of limited use
for a weapon meant to be used in mask range, and it's stated to not have
a string).
* Reverted the change to Magic Sign: Open Lock as making it a consumable
evidently now crashes it.
* Minor typofix.