Location updates
* Minor fix for a dresser that was meant to be a bed.
* Changed how beds and dressers spawn their items, creating more
reliable spawns that fit the theming better.
* Likewise made corpses spawn with slightly updated itemgroup for
clothing sets, streamlining the spawns a bit.
* Separated casualty spawns based on whether they were meant to have
died before the cataclysm or during it, with the older casualties only
spawning bones instead of a corpse. Effectively clarifies by implication
that the people in the curious structure were killed much more recently
than the other locations.
* Minor snippet updates for the strange grove note.
* Couple item placement fixes for the strange grove.
*