Assorted tweaks
* Changed which sprite active candles of warding use, added a variant
used in a specific context.
* Implemented investigate limit zone for the hermit. Doi.
* Added decorative candle pedestals to the rural church, used to hold
something placed after completing the first mission to light up the area
just a bit.
* Shifted the rural church over a bit as its proximity to the initial
fortifications was uneven, and that also meant the edges of the church
were right up against the farm's fencing. Moving it means I don't have
to take away more space from what'll eventually be a forge/workshop.
* Also slightly shifted tile position of the candle sprites to look
better when on top of the background for candle barriers, candle
pedestals, etc.