Start work on some of the Paragon of The Veil spells
* Finished up work on its version of Agility, implementing the intended
stats.
* Also added its versions of Clairvoyance and Confuse Monster.
* Implemented two new spells. Replacing Conjure Lightning Field and
Conjure Smoke are Radiant Burst and Conjure Shadow Field. The former is
a cone attack, lightning with a blinding effect. The latter creates a
newly-added field meant to black out the afflited area, with a side
effect of minor cold damage and more-lingering blindness for anything
you hit it with.
* Smol typo fixes.