Giant pile of assorted updates
* Added a fair bit of bufffer space surrounding the rural church, to
reduce the odds that it'll spawn grinding up against an anthill or
something else that would make it a stupid choice of holdout.
* Set it so that Urban Awakening is a lone start, no starter NPC. You've
been stuck down there alone, whereas with Seeker of The Arcane it's
implied you had chance to recruit assistance (plus or minus the current
inability to make starter NPCs have different dialogue).
* Forgot to set dialogue checks that look for Arcanist and Cleansing
Flame profession traits to also check for the variants Arcane Purifier
and Summoner get. Fixed that.
* Also added a bit more fleshing out of some dialogue. While I still
have plans to add more variation for Arcane Purifiers, Summoners now get
some expanded dialogue with some Cleansing Flame NPCs, reflecting their
status as the only non-faction arcanists mage hunters have a particular
beef with.
* Further dialouge expansions include allowing a non-sanguinist
Dragonblood to get flak for having used rather high-end blood magic to
end up the way they are, with ability to explain away that they have no
ties to the Sanguine Order. Paragons of The Veil can also get a bit more
interaction with some NPCs that might pick up on their otherwordly
connection, and summoners that trigger the same bad vibes can alleviate
the NPC's concerns as well, potentially getting a fair bit of coin if
they have a copy of To Summon The Unknown on hand and are willing to
turn it in. Speaking of which...
* Accordingly, altered the prices of arcane books in general. All of
them now have not only varied pre-apocalype prices based off how much
arcane potential offered by their teachings, but also post-apoc prices
based on how much the barrier to entry hinders applying that knowledge.
* Small price tweak to another item too. Oughta go back and revisit item
prices as a whole sometime.
* Added the light property to candle barriers and related items since
it's been an option for a while.
* Added more reminders about the whole "oi you gotta actually bind a key
for casting spells" thing in fitting places.
* Couple assorted typo fixes here and there.