Assorted smol things
* Renamed mana gem to essence gem, so players using Arcana with
Magiclysm won't confuse it with something linked to the mana system.
* Renamed spellcraft scrolls to pattern scrolls, to make their function
as bearing spell patterns clear, and to due to "spellcraft" being
associated with the Magiclysm skill.
* Some small rebalances and tweaks to assorted summoned monsters. In
particular, skeletons are a bit less flimsy, and all summoned monsters
get path settings and AI flags to make them less dumb.
* Made it so that obsolete spel-mutations are more clear about being
depreciated.
* Gave summoned shadow critters very low-volume, infrequent parrot
entries, to hint at something plot-wise...
* Updated some item descriptions that were based on older ideas and
references, and to reinforce some new plot concepts. In particular, one
of those tweaks ties in with not just another in this commit, but an
earlier thing set up...
* Couple typo fixes here and there.