Spell tweaks
* Change Light from a halo-creating spell to a field-creating one, with
a hidden field that exists only to emit light. PoTV version becomes able
to place location of the effect, chargen version is fixed at caster's
location.
* Change PoTV versions of Acid Armor, Flame Armor, Lightning Ward, and
Ward Against Evil to make use of immunity flag stuff. That last one,
PoTV's version being Holy Ward, is presently in a less than ideal state
as the up-and-coming PR only adds damagetype immunities for the moment,
and cut/bash/stab immunity is one hell of an effect to dole out.
* Changed Invisibility from using a halo item to the invisibility effect
flag, also from an upcoming PR. The effect that grants it also adds a
few other immunties and flavor effects. The resulting lack of
footstep-alteration is an acceptable balancing factor.
* Removed some item-hack stuff that's now obsolete.