Blood Magic Overhaul
* Implemented a ritual using the athame as a spell, that generates blood
essence when completed. Still creates blood loss at the end, but this is
now a side effect instead of part of the process. This makes spamming
rags no longer needed to generate blood essence via the athame.
* Likewise reworked the chalice to use this spell setup too, and made
the essence gain relative to the athame actually worth dealing damage.
As a side benefit, it'll actually mess with stamina like the action
message has always implied.
* Both spells also make use of the long-term side effects that I put
together in previous commits, so spamming these items constantly will be
problematic in the long run, even with magical healing/lifedrain to
offset the potential damage.
* Moved active athames, active chalice, and used chalice to
obsolete.json, along with obsoleting the recipes they played a part in.