Various updoots
* Set it so that Arcane Purifier and Summoner get their unique spells
from profession spells, instead of from their profession traits.
* Set it so that arcane blessings also start to supercede their
immediate magic sign counterparts, like they did in the past. Reduces
the eventual amount of spell clutter late in the game. It's less
agressive than the way Paragon of The Veil spells supercede Dragonblood
spells, kicking in when you develop towards the competing arcane
blessing instead of the first blessing in a competing mutation chain.
* Undid the scaling for Might, Agility, and Insight effects. Turns out
max_effective_intensity is broken to hell and back, so it would ramp up
stupid amounts of stat gains instead of capping out at +8 like it's
meant to.
* Expanded the available trait list for both scenarios, a bit.
* Prevent mage hunters from starting with Summon Skeletal Dog or
Serpentine Shield during chargen, as summoning spells are forbidden to
The Cleansing Flame.
* Also prevent mage hunters from starting with Dragonblood mutations in
chargen, as this typically involves blood magic.
* Prevent arcane purifiers from starting with Clarity of Blood during
chargen, as this qualiifes as blood magic.
* Prevent blood mages from starting with Healing during chargen, as
straight healing magic is contrary to the most common attitude towards
the Path to Power.
* Prevent keepers of the oath from starting with Dragonblood mutations,
as these explicitly tie users to the Beyond in ways forbidden by their
god.