Fix Dragonblood breaking Keeper of The Oath
Turns out that trying to mutate Dragonblood would make it eat the
profession trait keepers have. And since it turns out the chargen
restriction would normally fully subsume a mutation, I decided that it's
better to just let keepers be dergs if they really want, as only Paragon
of The Veil is intended to fully preclude any non-spell mutations.