Update essence gem
* Give essence gem a sensible volume, still enough to make storing large
amounts of essence worth it even in inventory.
* Allow storing crystallized essence in it.
* Add a draw cost. Turns out the impact of reloading from inventory,
even with something as small and light as essence, is a full turn at
minimum, but draw cost just outright replaces that. This gives essence
gems an actual use beyond being a workaround for dumb volume bugs.
* Updated item description to specifiy its added use, and point out you
have to be wearing it.

Closes https://github.com/chaosvolt/cdda-arcana-mod/issues/93