Updates to the hermit's missions
* Updated the hermit's missions to make use of MGOAL_CONDITON instead of
the earlier hack of using MGOAL_NULL.
* Implement some snaity-checking to the hermit for if the player is
attempting to tackle the final mission with a threshold that would
preclude becoming a Paragon of The Veil. Boom, fission mailed.
* Also made it so that screwing up while said final mission is in
progress, whether by becoming a post-thresh mutant in the process or by
simply dying, likewise updates the hermit so that he's properly aware
that you've either screwed up or died.
* Added a similiar failure condition if you screw up (presumably via
dying) the hermit's third mission.
* Added a backdoor for what happens if the player messes up the missions
to become a Paragon of The Veil, giving another character a shot at
accomplishing this. This includes the potential option of a mutie
character simply handing the sword over for safekeeping, so the next
character to earn his trust can go right back to eating up portals with
it.
* Also added a few mission descriptions along the way.