Further tweaks
* Changed speed tweaks to movecost, as the intended effect is weapon
momentum, not an all-around speed boost. Thanks to Hymore for revealing
that actually works.
* Made it so that Righteous Momentum only gets a max of -20 movecost at
max, and it decays faster so you have to attack constantly to build it
up.
* Shifted the bonus blocks and dodges to the two versions of Hunter's
Advance, fixed at 2 instead of being able to build up 4 like you
previously could.
* Reduced duration of both versions of Hunter's Advance to 2, making it
more liklely that you'll only proc the counterattack right at the
beginning or end of a fight.
* Nerfed the armor and armor penetration of the on-hit effects again, to
the functional mimimum.