Mod Interaction: The Bone Seer
* Characters with any of the notable arcanist traits (professions, as
well as Paragons of The Veil and Dragonbloods) get some added dialogue
after completing the bone seer's mission, which can lead to them
potentially offering an arcane monsterpart. Just a bit of flavor
dialogue that adds to their interaction, with the selection of parts
that get an interesting response being specifically ones used to create
a bone charm.
* Paragon of The Veil characters can muse on the nature of this implied
cycle when the subject of ending it comes up, leading to more flavor
dialogue between the two. This dialouge is also changed up a bit
depending on whether or not they've completed the seer's mission, and
whether or not they've given the seer a glimpse into the arcane.
* Mage hunters can respond with suspicion when the subject of
potentially destroying the world comes up, allowing them to respond with
hostility if they decide not to dismiss the seer's ideas as delusional.
* Mage hunters can also sidetrack the mission dialogue via accusing the
seer of practicing blood magic, which the seer responds to with
basically not really knowing what the hunter's on about, and
philosophizing about it instead. The player can either accept it and go
back to pondering the mission, or react with hostility.