Assorted ideas
* Overhauled the spear of pestilence. Now it gives a movecost bonus if
you're wielding it, and a chance of inflicting paralytic poison in
melee. In additon, instead of the default AEA_BUGS it instead spawns a
random number of summoned giant centipedes, with a random duration as
well. Using this effect is now less essence-intensive as well, though it
eats up plant life like Infernal Flame does.
* Gave active incorruptible swords and the restored ritual blade
NONCONDUCTIVE.
* Reduced active incorruptible sword's bonus cutting damage a bit,
increased its rate of energy consumption, and replaced VORPAL with a
crit technique that adds a dose of bonus fire damage, making it less of
an instant kill in exchange for triggering a bit more often.
* Implemented a proposed suggestion that gives the summoner profession
an air talisman and a makeshift short rope.