More stuff
* Set it so that current variants spawn a placeholder guard NPC where
the shopkeep is intended to be, and spawn the variant-specific
furnishings in one of the back rooms of the meeting hall.
* Added the furnishings and items that spawn when you finish the
mission, fleshing out three of the four rooms as barracks and making the
main section a proper dining hall with kitchen spaces.
* Re-implemented the "to be continued" check for now, pushed back so
that it triggers at the end of Sofia's fourth mission.
* Added some brief option dialogue for Sofia and Johnathan after
completing the fourth mission, flavor dialogue for getting their
opinions of
* Implemented variables and mapgen stuff at end of mission to make all
of the above function.
* Fixed a missing NPC-specific trait.