Mission stuff
* Implemented missions for the current crop of merchants you can attract
to the rural church. Each wants something in particular and pays you for
it, in addition to fleshing out the merchant's stock a bit more in some
way. Two versions also throw in free stuff of one sort or another in
addition to currency.
* Fixed up incorrect spawning of merhants, the doctor and hunter were
swapped.
* Fixed potions being dumped into Magiclysm's book itemsgroup instead of
potion itemgroup.
* Fixed merchant extra inventory for Wasteland Scavengers version of the
mission not counting as owned.
* Beefed up spawn chances of extra shop stuff.