Mutation additions/changes
* Gave Dragonblood another late-stage mutation, which makes use of scent
modifer traits so that they're harder for zombies and other otherworldly
monsters to detect, but easier for more earthly monsters to smell. Also
adds some crafting/reading penalties in exchange for use of the same
mutation flag as Apex Predator.
* Set it so that Paragons of The Veil go unnoticed by summoned monsters,
gain a minor craft bonus to arcana, but also draw increased hostility
from nether and fungal monsters.
* Implemented use of mutation copy-from to ensure that Cannibal and
Carnivore count as Dragonblood mutations, instead of relying on prereq
drag-in.
* Set it so that Metabolic Resilence is also needed for Dragonfire.
* Removed Animal Discord since it should now be redundant.
* Made it so that Curved Claws adds random damage to unarmed attacks
instead of granting a bonus attack.
* Added a species entry for Magiclysm dragons, as a
mod-intercompatability thing for late-stage Dragonbloods.