Ranged weapon updates
* Gave the symbol of judgement a custom ammo effect instead of relying
on existing ones, giving its AoE lightning a wider, more erratic
placement.
* Added a low chance of spawning anomalous fire to the demon claw's ammo
effects.
* Implemented the currently-pending improvement that allows
explicitly-set ranged damage to have armor pierce.
* Tweaked wraithslayer crossbow's damage type now that stab is becoming
the norm for ranged damage instead of cut. Use of "true" damage means
that its armor-pierce is effectively irrelevent, but the pierce is still
defined to better convey to the player that it's effective against armor
(a problem that armor_multiplier had).
* Updated item descriptions for a couple of said guns.