Crafting updates
* Changed the silver athame to use either blood essence or a flame
talisman instead of needing both blood essence AND the materials
formerly required to make blood essence. Aside from eliminating a
hard-to-acquire reference to a now obsolete recipe, it also changes the
power source to potential be easier to acquire, while the flame
talisman's also reference the current common ways to produce blood
essence.
* Set it so blood-soaked rags can still, if desired, be used instead of
fur or leather for the athame's handle, and added a need for a prior
source of cutting quality for said handle wrap (same as veinreaver).
* Also changed it so veinreaver can use leather or fur as alternatives
to blood-soaked rags for the handle wrap, as above.
* Changed the hammer of the hunter to allow 1 dull essence or a water
talisman instead of 1 essence, again offering an alternative that
references the method of generating dull essence.
* Changed the offering chalice to need either 1 essence or an earth
chalice to craft, instead of previous essence demands. Makes it
potentially easier to acquire depending on available resources, and the
recipe for earth talismans reference product of normal essence.