Recipe variants
Added some recipe variations for specific booklearn-only equipment that
faction arcanists use to make their specific magic items, making it
possible to craft them using the faction arcane books instead of the
medieval smithing books.

General theme is using half the fuel, 75% the time, and one less
fabrication skill to craft them. However, in exchange it requires equal
skill in arcana and a different toolset than normal forging. Rather than
using a swage it requires a crucible (where not already used), the
presence of a candle of warding, and the addition of an earth talisman
and some essence.