misc improvements
* Swapped out the bar doors in the sanguinist holdout for pickable metal
doors, so they're less likely to clash with the captives.
* Set it so that the host of the archon is no longer friendly to
vortices for lore reasons, but also gains a spell effect in his special
attack that destroys vortices outright.
* Made summoned vortices no longer electric, negating their immunity to
the corrupted symbol of judgement the host of the archon uses.
* Updated Dragonblood's use of anger_relations to account for how low
the aggression needs to be to work right.
* Fixed the recruitable sanguine apparentice and shrike lacking their
background story traits.
* Replaced the hermit's symbol of judgement with a random, more mundane
weapon to not make it as blatant that he's amarcanist.