Add Sofia's penultimate mission
* Added a small hideout of renegade sanguine shrikes, who Sofia sends
you to come into contact with. Talking them into giving you want you're
after is an option, albeit this is much harder as a mage hunter
character.
* Main quest target can be persuaded via a few different ways, and
additionally by talked into trading with you if you want.
* Completing this mission peacefully also spawns another side NPC at the
rural church, a minor merchant and likely to be good for sideplot stuff
later on.