Large assortment of misc updates
* Made dull essence more granular. Now there are 30 dull essence per
* Updated recipes, item spawns, etc with the above change in mind.
* Updated items that use dull essence, some have slightly different
amounts of max charge, and some were changed to be more or less
efficient, instead of just doubling max charges and halving charges per
turn.
* The above changes were also to support rebalancing of magic ranged
weapons. With the exception of the demon claw (which has an explosion
compensating for its reduced efficiency in direct damage) and bane staff
(which is MORE damaging relative to ammo usage), it's now expected that
1 essence will equal 60 damage. Now that 1 essence equals 30 dull
essence as well, set the wraithslayer to use 20 dull essence per shot
and deal 40 damage.
* Fixed mi-go evolutions not having their drop overrides.
* Gave various cult and nether monsters classic category, because if
you're comining Arcana with Dark Days of The Dead you're presumably
WANTING weird monstrs to still spawn.
* The more lovecraftian nether monsters received the COLDPROOF flag,
along with the spirit of fire and host of the archon.
* Tweaked part of Sofia's dialogue so it checks for if the player has
basic-level arcana skill.
* Gave essences a token level of volume. This also means that the
effective energy density of essence changes with each stage, reducing
the impact that "forced minimum volume" has on the question of each
essence type is better for storage.
* Set magic seals to cast at level 5 instead of 10.
* Changed Destructive Absorption to use cold-type damage, and increased
the chance that it'll spawn dimensional fatigue.
* Added mod_tileset sprites for air talismans, the new boss, and the new
traps. Also an updated version of the host of the archon's sprite,
thanks to help from SDG.