Workaround for broken spellcasting comestibles
* Re-implemented the ancient hack I had to use for magical scrolls,
because comestibles with the cast_spell use action are broken. No fix
seems imminent so far despite it also affecting Magiclysm, so we're back
to transforming tools to fake consumables.
* Also implemented it for magic seals. I don't recall if having it
vanish on use works properly anymore or if it now needs a revert,
meaning...
* Added a material for magic seals to revert to when used up, nothing
more than a clump of ash.
* Also fixed red potion being broken due to it having relied on casting
a spell to grant a little bonus healing. Now it's back to only giving
the elixir buff.

Still need to figure out what to do with air talismans, as now they're
not quite useful since glyph summons count as pets...

Relevent issue:
https://github.com/CleverRaven/Cataclysm-DDA/issues/41482