Preparation for removal of artifacts
* Updated the clairvoyance halo effect to use the new mutation-based
method, and reduced it to basic-ranged clairvoyance.
* Changed how Serpentine Shield works. Basically, on-hit relic stuff
can't handle summoning monsters as an effect, so its replacement now
blinds and zaps the attacker.
* Replaced Dampening outright. Not only is AEP_EXTINGUISH not really
something you can recreate with relics, even then it was a shitty spell
that wasn't really worth spending chargen points on. Replaced it with a
`SWAP_POS`-based spell designed for a quick entrance or exit to combat.
* Since Dragonblood previous got a slightly-upgraded version of
Dampening, gave them an upgraded, more lethal take on the new spell.
* Set the life-drain horned halo to use a new on-hit healing effect. One
obvious issue with it: `hit_you_effect` stuff triggers even when you
miss, so it's kinda silly. As close to functional as this is going to
get, I suppose.
* Set it up so that the demon claw has the new life drain (at less than
100% rate via `once_in`), and a custom take on AEP_EVIL. In a nutshell,
the strength boost that's supposed to offset the debuffs from Eeeeviiil
is a hardcoded result of items with AEP_EVIL, getting tagged with the
effect itself isn't good enough. So instead I added an effect that gives
a decent shifting stat setup that makes it worth trying to use the item
at first, but slowly deteriorates if allowed to progress.
* Reworked the mask of insight as best as I could. Needless to say,
relics being not fully functional mandated a massive shift in how the
item works. The best way to preserve its functionality proved to be via
consuming a small burst of dull essence to grant a fake "glowing eyes"
sorta effect similiar to the halo items, which gives you all the needed
side effects of the mask. Only real plus side to this extensive hackery
needed to regain lost functionality was the damn thing finally gives you
the sorta of "combat clairvoyance" dodge bonus I'd planned it to have.
* Hit the mantle of shadows with the same rework, making it spend a bit
of essence on a modest burst of invisibility. No longer muffles your
footsteps since that's not actually a thing relics can do yet. Instead
just grants an invisibility effect that also halts a few specific
effects, being a lesser version of Illusion Cloak.
* Removed the silly AEP_EXTINGUISH effect from the jade wreath. Doesn't
really warrant replacing it with anything.
* Reworked the meteoric talisman to give you a burst of a protective
effect instead of using transform hacks, as you'd likely expect by that.
Also removed the bad weather effect stuff. This also led to...
* Rebalanced the three protective lessor relics. Now they all give the
Greater version of their ward effect, and each use one essence's worth
of energy for 5 minutes worth of the buff, instead of varying amounts of
energy for 100 minutes.
* Removed AEP_PSYSHIELD and AEP_STR_DOWN from the ring of the fallen
angel. Since AEP_PSYSHIELD is a hardcoded thing that relics can't
replicate just yet, its main function will be its granting its
protective effect. Now however, like the lesser unique artifacts, it
grants full-power Holy Ward instead of Ward Against Evil. As a
consequence it's much less cost-efficient that the lesser uniques.
* Reworked the enchantments on the orb of the veil. Still gets +4
intelligence but now only +2 perception. Traded AEP_ATTENTION for its
mutation-based enchantment granting the full effects of Schizophrenic,
on top of downgrading its clairvoyance.
* Removed the electric resistance and bad weather effect on the meteoric
verge for now. Again, no real way to recreate AEP_BAD_WEATHER just yet,
and recreating electric resistance on this item also isn't really doable
sanely in the way it was implemented (I'd need it to basically spam an
intermittent spell that tags the user with immunity to electricity).
Instead I set it so that casting its spell gives the caster a few
seconds of Greater Lightning Ward.
* Reworked the moonstone fang's effects to include infrequent lifedrain,
a less-commonly-procced version of Shadowy Shield's effect, in exchange
for intermittent hits of the Touched Mind effect and healthiness
debuffs.
* Converted the restored ritual blade's intelligence boost to be
relic-based. And sadly, had to drop ARTC_PORTAL, not that it was working
anyway since artifact charging was broken by nested containers anyway.
* Converted the cursed blade to use the same item-spawning hack as the
mask of insight. In this case it uses its full proper duration instead
of short bursts like the other reworked items, since a lot of evil (or
in this case, the new pseudo-evil effect) buildup is part of the
intended side effects. Also tacked on Killer Drive as a bonus
enchantment mutation (will test if this works sanely on Pacifist
characters when enchantment mutations gets merged), using the sword's
spell pings your healthiness heavily, but the life drain effect is
stronger than other sources of lifedrain in arcana.
* Updated mutation and item descriptions for the affected content.
* Belatedly updated the martial arts file since some active items have
since been obsoleted.
* Couple misc effect updates and fixes.

Caused by https://github.com/CleverRaven/Cataclysm-DDA/pull/42203
hitting way earlier than I'd expected. I was hoping that the last 2-3
major overhauls would've been close to finished before they made a move
on artifacts...