Assorted fixes and updates
* Updated stat-boosting enchantment effects now that
https://github.com/CleverRaven/Cataclysm-DDA/pull/42700 fixed
stat-boost-doubling.
* Belatedly updated some mapgen variants that were using terrain that's
now furniture.
* Updated recipes to not use now-removed generic acids.
* Set it so Ward Against Poison protects against mi-go atmosphere.
* Set it so that Revenant Blood prevents you from feeling full, since
supernatural hunger is part of the side effects.
* Upgraded Revenant Blood to actively remove the effects it protects
against since it's meant to be an upgrade from Revenant Hunger AND
Clarity of Blood, and added the same fullness-effect-removal Revenant
Blood now gets.
* Gave Elemental Affinity an effect that makes protects Dragonblood
characters from the main effects of warmth. Went with still leaving them
vulnerable to heat slowdown and blistering, for now.