On-hit fixes, item rebalances/fixes, relic uncrafts
* Satchel of eternity now has its weight-saving restored via a new
method, no longer expands to take up wacky levels of volume when full of
stuff, and slight nerf to maximum content weight.
* Fixed on-hit effects for satchel of eternity, symbol of judgment,
hellfire staff, spear of pestilence, meteoric verge,
* Added a proc message to Shadowy Shield on-hit effects.
* Fixed proc message of bane staff.
* Removed the non-function on-hit effect from veinraver since bleeding
is no longer all that special, plus it doesn't actualy do its AoE like I
wanted it to.
* Removed the drain-life effect from moonstone scourge since
transforming relics don't work, it should only be gaining life-drain
when active.
* Removed the old broken life-drain on-hit effect from the cursed sword,
since it now gets it from the berserk aura effect given when activated.
* Updated affected_body_parts now that it's lowercase like all other BP
IDs.
* Beefed up the minimum damage of spell effects for meteoric verge,
wyrmskin armor based off the math that went into the sort of spell it'd
be equivalent to.
* Under the same logic, lowered healing for gilded aegis and jade
hauberk.
* Lowered max damage but raised min damage for scroll of sundering based
off some spell-balance thoughts, and swapped fire for toxic gas. Also
made the spike AoE add another status effect to targets.
* Tweaked damage range of thunder gauntlet effects as well.
* Minor rebalancing of some consecration recipes.
* Added a recipe to sacrifice the new items used as the base for vanilla
relics, gaining crystallized essence as a result. This means that Arcana
now officially has a way to make use of vanilla artifacts. Three
crystallized essence is a lot, but with how rare artifacts are supposed
to be it should be reasonable.