Spell updates, purge obsolete.json
* Updated spells to have shapes as warranted by
https://github.com/CleverRaven/Cataclysm-DDA/pull/44481. Would've saved
a lot of excess JSON to just assume "blast" to be a default, but oh
well. Now I have suffered as Korg suffered.
* Converted all uses of line, cone, etc attacks to use attacks with the
correct shape.
* Dumped in `NO_PROJECTILE` into appropriate places. There are PROBABLY
more places that could use it but 90% of the time there's no effective
difference between a projectile spell and a non-projectile spell since
you can't actually miss.
* Also, a fine time to cleanse this world of obsolete.,json since it had
spells in it that likely have not been on anyone's save file in ages.