Various fixes
* Update starting ammo for a couple professions.
* Update proficiencies for the more advanced professions.
* Fix Arcane Purifier having obsoleted lockpick skill instead of
trapping.
* Fixed purifer and shrike having holsters unsuited for their ranged
weapons.
* Removed storage pockets from wyrmskin armor and mantle of shadows,
because it turns out nested container stuff still fucks with tool-armor
being able to properly have both ammo and pockets.
* Updated item lengths for tools and ranged weapons, generally based off
closest equivalent item where possible.
* Gave the cursed blade and restored ritual blade volume, weight, and
length in between the vanilla longsword and zweihander, closer to the
lower bound for most historical zweihanders, so that it'll just just
barely fit in a back scabbard. Cursed blade still spawns outside its
back scabbard in the floating temple however, due to implying the
trapped Keeper in that room commited suicide with it (generally
difficult to do with a scabbard in the way).

Closes https://github.com/chaosvolt/cdda-arcana-mod/issues/163
Closes https://github.com/chaosvolt/cdda-arcana-mod/issues/162