Locations and spell updates
* Switched Magic Sign mutations back to not having upgrades again. This
is acceptable now that NO_FAIL allows for tinmkering with experience
granted per cast, so cast-grinding is an actual option now. As a side
benefit, this enables...
* Set it so that the arcanist scenarios can learn some of the magic
signs you'd normally need a pattern scroll for. The above is needed for
that so that selecting and unselecting that trait works correctly.
* Tweaked the spell selection so that the spell lists are a bit more
unique between Seeker of The Arcane and Urban Awakening. Might, Agility,
Insight, and Light are the only ones the two scenarios have in common
now, the remaining spell choices for each are unique to that scenario.
* Added arcane variants to libraries, house_library, churches, and
museums.
* Couple minor additions of arcana content to some newish vanilla
itemgroups.