Various minor updates
* Added silver potion, which requires combining alchemy with the secrets
in Oaths To The Chalice to make. Basically a mana potion.
* Fixed redundant demand of a bowl for making essence using an offering
chalice.
* Set it so crystallized essence doesn't require having the books on
hand as tools to make.
* Set the BN version of the restored ritual blade to have a second use
action that exists just to work around with bugs and regain any
crystallized essence loaded into it.
* Set it so you can unload the DDA version of the restored ritual blade,
since that's not needed without ARTC_PORTAL.
* Set it to spawn crystallized essence in the anomalies the hermit sends
you after in the DDA version, since ARTC_PORTAL is no longer a thing.
* Tweaked sizes and values of potions.
* Fixed potion liquid demand being all fucky.