Fix Arcana and Magicylsm intercompability issue
* Set it so that Dragonblood mutation automatically grabs Carnivore and
Cannibal as prerequisites of Inner Fire and Elemental Affinity, without
having to edit those mutations.
* Setting it so Carnivore is a prereq of Inner Fire instead of
Dragonfire prevents a potential segfault when using blood effigy.
* Removed the seraphic shade's ability to inflict psychiatric mutations
on the player for now, since a MUTATE_TRAIT wonder spell version of this
feature would be a lot more work for something that only has a 5% chance
of triggering, from an attack used by a boss that endgame characters
will encounter a grand total of once. Not really worth it, and the
encounter has other features to make the fight interesting and
Lovecraftian already.

This should fix the issue of Manatouched threshold, and possibly Black
Dragon threshold, being unobtainable when playing with both Arcana and
Magiclysm. Tested via loading a save with both, spawning in with an
assortment of traits, and chugging concentrated mana serum. It still
takes a while, but it's definitely not anything on Arcana's end now.

Having a bunch of starting trait put my mutation strengths at 10 for
Rat, 14 for Fish and Troglobite, 15 for Insect, 16 for Mouse, and 18 for
Elf-a, etc. No mutation strength for Paragon of The Veil or Dragonblood.
Manatouched's mutation selection is presumably just that poor. It's now
52 after breaching the threshold.

With thanks to @KorGgenT for pointing this out.